// src/app/core/api/generated/model/userRegisterDTO.ts
/**
 * VetClinicAPI_V2 | v1
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface UserRegisterDTO { 
    name: string;
    phone: string;
    email: string;
    licenceNumber?: string | null;
    password: string;
    role: number;
}