// src/app/core/api/generated/model/userResponseDTO.ts
/**
 * VetClinicAPI_V2 | v1
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { GetAnimalByIdIdParameter } from './getAnimalByIdIdParameter';


export interface UserResponseDTO { 
    id?: GetAnimalByIdIdParameter;
    name?: string;
    phone?: string;
    email?: string;
    licenceNumber?: string | null;
    role?: number;
    isActive?: boolean;
}